Module tendermock::avl

AVL Tree

This module hosts a simple implementation of an AVL Merkle Tree that support the get and insert instructions (no delete yet, it's not needed as the on-chain store is supposed to be immutable).

Proof of existence are supported using ICS23, but proof of non-existence are not yet implemented.

Keys needs to implement Ord and AsBytes (see as_bytes module), while values are required to implement Borrow<[u8]>.

For more info, see AVL Tree on wikipedia,

Re-exports

pub use as_bytes::AsBytes;

pub use proof::get_proof_spec;

Modules

as_bytes
AsBytes trait definition

proof
ICS23 Proof

tests
Test suite of tendermock AVL Tree.

Structs

AvlNode
A node in the AVL Tree.

AvlTree
An AVL Tree that supports get and insert operation and can be used to prove existence of a given key-value couple.

Constants

HASH_ALGO

Functions

as_node_ref
Wrap a key + value couple into a NodeRef.

Type Definitions

NodeRef