Module tendermock::avl::as_bytes

AsBytes trait definition

This module hosts the AsBytes traits, which is used by the AVL Tree to convert value to raw bytes. This is helpful for making the AVL Tree generic over a wide range of data types for its keys (the values still need to implement `Borrow<[u8]>), as long as they can be interpreted as a slice of bytes.

To add support for a new type in the AVL Tree, simply implement the AsByte trait for that type.

Traits

AsBytes
A trait for objects that can be interpreted as a slice of bytes.