Struct tendermock::avl::AvlTree

pub struct AvlTree<K: Ord + AsBytes, V> {
    root: Option<Box<AvlNode<K, V>>>,
}

An AVL Tree that supports get and insert operation and can be used to prove existence of a given key-value couple.

Fields
root: Option<Box<AvlNode<K, V>>>
Implementations

impl<K: Ord + AsBytes, V> AvlTree<K, V> where
V: Borrow<[u8]>,

pub fn new() -> Self

Return an empty AVL tree.

pub fn root_hash(&self) -> Option<&Hash>

Return the hash of the merkle tree root, if it has at least one node.

pub fn get<Q: ?Sized>(&self, key: &Q) -> Option<&V> where
K: Borrow<Q>,
Q: Ord,

Return the value corresponding to the key, if it exists.

pub fn insert(&mut self, key: K, value: V)

Insert a value into the AVL tree, this operation runs in amortized O(log(n)).

fn insert_rec(node_ref: &mut Option<Box<AvlNode<K, V>>>, key: K, value: V)

Insert a value in the tree.

pub fn get_proof<Q: ?Sized>(&self, key: &Q) -> Option<CommitmentProof> where
K: Borrow<Q>,
Q: Ord,

Return an existence proof for the given element, if it exists.

fn get_proof_rec<Q: ?Sized>(
&self,
key: &Q,
node: &Option<Box<AvlNode<K, V>>>
) -> Option<ExistenceProof> where
K: Borrow<Q>,
Q: Ord,

Recursively build a proof of existence for the desired value.

fn balance_node(node_ref: &mut Option<Box<AvlNode<K, V>>>)

Rebalance the AVL tree by performing rotations, if needed.

fn rotate_right(root: &mut Option<Box<AvlNode<K, V>>>)

Performs a right rotation.

fn rotate_left(root: &mut Option<Box<AvlNode<K, V>>>)

Perform a left rotation.