Module tendermock::grpc

gRPC

The gRPC interface of tendermock, for now most of the queries are unimplemented.

The serialization is handled by prost, a gRPC framework that generates all the desialization/deserialization code from protobuf files. The protobuf files and generated Rust code lives in the ibc_proto crate.

The server code is also generated, this time by tonic and it also lives in the ibc_proto crate. This module simply implements the Query trait generated by Tonic on a custom QueryService struct.

Structs

QueryService
A struct handling the Query service.

Functions

new
Create a new gRPC server.

Type Definitions

GrpcServer