Module tendermock::jrpc

JsonRPC

This module contains the the utilities needed do build the JsonRPC api, both HTTP and WebSockets, which can then directly served by warp.

warp is a HTTP framework which is built around filters, this module is used to build filters that mimics the Tendermint API.

Re-exports

pub use api::Jrpc;

pub use websockets::Ws;

Modules

api
The Tendermock JsonRPC HTTP API.

utils
A JsonRPC mini-framework.

websockets
The Tendermock JsonRPC Websocket API.

Constants

WEBSOCKET_PATH

Functions

serve
Create a new gRPC server.