Module tendermock::jrpc::utils

A JsonRPC mini-framework.

This module provides a small framework for building JsonRPC API on top of a wrap filter, this is done through the builder methods of JrpcFilter.

Structs

JrpcCtx
JsonRPC context.

JrpcEnvelope
JsonRPC envelope.

JrpcErrorDetails
JsonRPC error details.

JrpcFilter
A JsonRPC warp filter, methods can be added through the builder method add while the filter is built with the build method.

JrpcResponse
JsonRPC response.

Enums

JrpcError
JsonRPC errors.

Constants

JRPC_CODE_INVALID_PARAMS

JRPC_CODE_INVALID_REQUEST

JRPC_CODE_METHOD_NOT_FOUND

JRPC_CODE_SERVER_ERROR

JRPC_VERSION

Type Definitions

JrpcResult

MethodMap

SharedMethodMap