Module tendermock::node

Node

A tendermock node, which encapsulate a storage and a chain. This also exposes a thread safe version called SharedNode for use by the various RPC interfaces.

To integrate with IBC modules, the node implements the ICS26Context traits, which mainly deal with interacting with storing and reading values.

Structs

Connections
A type representing connections in memory

Node
A node contains a store, a chain and some meta-data.

SharedNode
An Arc<RwLock<>> wrapper around a Node.

Constants

CLIENT_STATE_URL

COMMITMENT_PREFIX

CONSENSUS_STATE_URL