Module tendermock::store

Store

A storage for tendermock. For now the only available storage is the InMemoryStore, which as its name imply is not persisted to the hard drive. However, possible implementation of persistent storage are possible without impacting the rest of the code base as it only relies on the Storage trait, which may be implemented for new kinds of storage in the future.

Structs

InMemoryStore
An in-memory store backed by a simple hashmap.

Traits

Storage
A concurrent, on chain storage using interior mutability.